// src/types/department.ts
export type Department = {
  id: number
  sku_corporation?: string
  sku_corporation_child?: string
  sku_department?: string
  sku_department_parent?: string
  name: string
  route?: string
  icon?: string
  modules?: string
  description?: string
  status?: string
  created_at: string
  updated_at: string
}